+++ src/types/styles/index.ts
@@ -12,3 +12,13 @@ export const HostDisplayBlockStyle = ':host{display: block;}';
  * Css display flex style to host of component
  */
 export const HostDisplayFlexStyle = ':host{display: flex;}';
+
+/**
+ * Css display flex and flex direction column style to host of component
+ */
+export const HostFlexColumnStyle = ':host{display: flex;flex-direction: column;}';
+
+/**
+ * Css display flex and flex direction row style to host of component
+ */
+export const HostFlexRowStyle = ':host{display: flex;flex-direction: row;}';