+++ structured-log/src/index.ts
@@ -9,3 +9,5 @@ export * from './services/console/consoleSinkConfig.service';
 export * from './services/logger.service';
 export * from './services/rest/restSinkConfig.service';
 export * from './services/rest/restSink.service';
+export * from './services/developerConsole/developerConsoleSink.options';
+export * from './services/developerConsole/developerConsoleSink.service';